+++ ReleaseNotes.md
@@ -4,11 +4,21 @@
 
 ## Breaking changes
 
-* TGR-1362: HttpGlueCode now waits by default for the request to be received before continuing. This can be changed by setting the configuration key `tiger.httpClient.executeBlocking` to `false`.
+* TGR-1362: HttpGlueCode now waits by default for the request to be received before continuing. This can be changed by setting the configuration
+  key `tiger.httpClient.executeBlocking` to `false`.
 
 ## Features
 
 * TGR-1372: Zion: The responseCode is now a string, meaning it now supports the use of TigerConfiguration values.
+* TGR-1377: Tiger-Test-Lib:
+    * Added Request Validation Steps:
+        * @Then("TGR current request body matches:")
+        * @Then("TGR current request with attribute {tigerResolvedString} matches {tigerResolvedString}")
+        * @Then("TGR current request contains node {tigerResolvedString}")
+        * @Then("TGR current request at {tigerResolvedString} matches:")
+        * @Then("TGR current request at {tigerResolvedString} matches as {modeType}:")
+        * @Then("TGR current request with attribute {tigerResolvedString} does not match {tigerResolvedString}")
+    * Updated user manual, see section "Tiger Test Lib > Validating requests"
 
 ## Bugfixes